+++ tests/test_brute_force.py
@@ -1,5 +1,3 @@
-from __future__ import division, print_function, absolute_import
-
 from numpy.testing import (
     assert_equal,
     assert_
@@ -8,119 +6,6 @@ from numpy.testing import (
 import numpy as np
 from grispy import GriSPy
 import pytest
-
-
-class Test_data_consistency:
-    @pytest.fixture
-    def setUp(self):
-
-        np.random.seed(1234)
-        self.centres = np.random.rand(3, 5).T
-
-        self.data = np.array(
-            [
-                [0, 0, 0],
-                [0, 0, 1],
-                [0, 1, 0],
-                [0, 1, 1],
-                [1, 0, 0],
-                [1, 0, 1],
-                [1, 1, 0],
-                [1, 1, 1],
-            ]
-        )
-
-        self.upper_radii = 0.7
-        self.lower_radii = 0.5
-        self.n_nearest = 5
-
-        self.gsp = GriSPy(self.data)
-
-    def test_grid_data(self, setUp):
-        assert_(isinstance(self.gsp.dim, int))
-        assert_(isinstance(self.gsp.data, np.ndarray))
-        assert_(isinstance(self.gsp.k_bins, np.ndarray))
-        assert_(isinstance(self.gsp.metric, str))
-        assert_(isinstance(self.gsp.N_cells, int))
-        assert_(isinstance(self.gsp.grid, dict))
-        assert_(isinstance(self.gsp.periodic, dict))
-        assert_(isinstance(self.gsp.periodic_flag, bool))
-        assert_(isinstance(self.gsp.time, dict))
-
-    def test_bubble_single_query(self, setUp):
-
-        b, ind = self.gsp.bubble_neighbors(
-            np.array([[0, 0, 0]]), distance_upper_bound=self.upper_radii
-        )
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), 1)
-        assert_equal(len(ind), 1)
-
-    def test_shell_single_query(self, setUp):
-
-        b, ind = self.gsp.shell_neighbors(
-            np.array([[0, 0, 0]]),
-            distance_lower_bound=self.lower_radii,
-            distance_upper_bound=self.upper_radii,
-        )
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), 1)
-        assert_equal(len(ind), 1)
-
-    def test_nearest_neighbors_single_query(self, setUp):
-
-        b, ind = self.gsp.nearest_neighbors(
-            np.array([[0, 0, 0]]), n=self.n_nearest
-        )
-
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), 1)
-        assert_equal(len(ind), 1)
-        assert_equal(np.shape(b[0]), (self.n_nearest,))
-        assert_equal(np.shape(ind[0]), (self.n_nearest,))
-
-    def test_bubble_multiple_query(self, setUp):
-
-        b, ind = self.gsp.bubble_neighbors(
-            self.centres, distance_upper_bound=self.upper_radii
-        )
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), len(self.centres))
-        assert_equal(len(ind), len(self.centres))
-
-    def test_shell_multiple_query(self, setUp):
-
-        b, ind = self.gsp.shell_neighbors(
-            self.centres,
-            distance_lower_bound=self.lower_radii,
-            distance_upper_bound=self.upper_radii,
-        )
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), len(self.centres))
-        assert_equal(len(ind), len(self.centres))
-
-    def test_nearest_neighbors_multiple_query(self, setUp):
-
-        b, ind = self.gsp.nearest_neighbors(self.centres, n=self.n_nearest)
-        assert_(isinstance(b, list))
-        assert_(isinstance(ind, list))
-        assert_equal(len(b), len(ind))
-        assert_equal(len(b), len(self.centres))
-        assert_equal(len(ind), len(self.centres))
-        for i in range(len(b)):
-            assert_equal(np.shape(b[i]), (self.n_nearest,))
-            assert_equal(np.shape(ind[i]), (self.n_nearest,))
-
 
 class Test_grispy:
     @pytest.fixture
@@ -138,7 +23,6 @@ class Test_grispy:
         self.n_nearest = 32
         self.eps = 1e-6
 
-        self.gsp = GriSPy(self.data)
         periodic = {0: (-self.lbox * 0.5, self.lbox * 0.5)}
         self.gsp = GriSPy(self.data, periodic=periodic)
 
@@ -232,7 +116,7 @@ class Test_grispy:
             assert_equal(len(b[i]), len(d))
             np.testing.assert_almost_equal(b[i], sorted(d), decimal=14)
 
-    def test_nearest_newighbors_precision(self, setUp_1d):
+    def test_nearest_neighbors_precision(self, setUp_1d):
 
         b, ind = self.gsp.nearest_neighbors(self.centres, n=self.n_nearest)
 
@@ -249,3 +133,99 @@ class Test_grispy:
             d = d[: self.n_nearest]
             assert_equal(len(b[i]), len(d))
             np.testing.assert_almost_equal(b[i], d, decimal=16)
+
+class Test_periodicity_grispy:
+    @pytest.fixture
+    def setUp_1d(self):
+
+        self.lbox = 10.0
+        self.data = np.array(
+            [
+                [2, 0, 0],
+                [-2, 0, 0],
+                [0, 2, 0],
+                [0, -2, 0],
+                [0, 0, 2],
+                [0, 0, -2]
+            ]
+        )
+
+        self.eps = 1e-6
+        self.gsp = GriSPy(self.data)
+
+    def test_periodicity_in_shell(self, setUp_1d):
+
+        centres = np.array([[0.0, 0.0, 0.0]])
+        upper_radii = 0.8 * self.lbox
+        lower_radii = 0.8 * self.lbox
+
+        for j in range(3):
+            self.gsp.set_periodicity({j: (-0.5 * self.lbox, 0.5 * self.lbox)})
+
+            dis, ind = self.gsp.shell_neighbors(
+                centres,
+                distance_lower_bound=lower_radii - self.eps,
+                distance_upper_bound=upper_radii + self.eps,
+            )
+            dis, ind = dis[0], ind[0]
+
+            aux = np.argsort(ind)
+            ind = ind[aux]
+            dis = dis[aux]
+            for i in range(2):
+                assert_equal(ind[i], i + (j * 2))
+                assert_(dis[i] >= lower_radii - self.eps)
+                assert_(dis[i] <= upper_radii + self.eps)
+
+        self.gsp.set_periodicity(
+            {
+                0: (-self.lbox * 0.5, self.lbox * 0.5),
+                1: (-self.lbox * 0.5, self.lbox * 0.5),
+                2: (-self.lbox * 0.5, self.lbox * 0.5),
+            }
+        )
+        dis, ind = self.gsp.shell_neighbors(
+            centres,
+            distance_lower_bound=lower_radii - self.eps,
+            distance_upper_bound=upper_radii + self.eps
+        )
+        dis, ind = dis[0], ind[0]
+
+        aux = np.argsort(ind)
+        ind = ind[aux]
+        dis = dis[aux]
+
+        for i in range(6):
+            assert_equal(ind[i], i)
+            assert_(dis[i] >= lower_radii * (1.0 - self.eps))
+            assert_(dis[i] <= upper_radii * (1.0 + self.eps))
+
+    def test_periodicity_in_bubble(self, setUp_1d):
+
+        centres = np.array(
+            [
+                [5.0, 0.0, 0.0],
+                [0.0, 5.0, 0.0],
+                [0.0, 0.0, 5.0]
+            ]
+        )
+        upper_radii = 0.3 * self.lbox
+
+        for j in range(3):
+            self.gsp.set_periodicity({j: (-self.lbox * 0.5, self.lbox * 0.5)})
+
+            centre = centres[j].reshape(1, 3)
+
+            dis, ind = self.gsp.bubble_neighbors(
+                centre, distance_upper_bound=upper_radii
+            )
+            dis, ind = dis[0], ind[0]
+
+            aux = np.argsort(ind)
+            ind = ind[aux]
+            dis = dis[aux]
+
+            for i in range(2):
+                assert_equal(ind[i], i + (j * 2))
+                assert_(dis[i] <= upper_radii * (1.0 + self.eps))
+                assert_(dis[i] >= upper_radii * (1.0 - self.eps))
